GCP cloud function/requirements.txt
# Function dependencies, for example:
# package>=version
joblib
google-cloud-storage
sklearn